feat: Refactor ProductCard and create UpdateModal component
